requirements.txt
pyserial
argparse
Flask
gevent
